refactor: extract types in type file
